Cleanup, some error handling, and UI updates
